requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.13
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
docx2txt==0.8
durationpy==0.9
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.2
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.35.0
googleapis-common-protos==1.65.0
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.6
httptools==0.6.2
httpx==0.27.2
huggingface-hub==0.25.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kubernetes==31.0.0
langchain==0.3.3
langchain-chroma==0.1.4
langchain-community==0.3.2
langchain-core==0.3.10
langchain-openai==0.2.2
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.0
langchain-upstage==0.3.0
langsmith==0.1.134
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==2.9.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.51.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.3.3
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
posthog==3.7.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.37.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.28.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.40.0
streamlit==1.50.0
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
yarl==1.15.2
zipp==3.20.2
openpyxl==3.1.5